Attempted on enemy physX

+++ Final project game thing/src/Gun.java
@@ -7,6 +7,16 @@ public class Gun extends Weapon
 	private ArrayList<Bullet> bullets;
 	private double bulletSpeed;
 	
+	/**
+	 * Creates a custom Gun object that deals certain damage, can fire
+	 * upto a particular range, shoots bullets at a rate, with a unique ID
+	 * @param damage the damage the weapon can deal
+	 * @param range the range of the weapon
+	 * @param bulletCount the amount of bullets held by the gun
+	 * @param fireRate the rate of fire 
+	 * @param speed the speed of the weapon ???????
+	 * @param ID the ID of the weapon
+	 */
 	public Gun(double damage, int range, int bulletCount, double fireRate, double speed, int ID)
 	{		
 		super(damage, fireRate, ID);
@@ -15,10 +25,24 @@ public class Gun extends Weapon
 		this.bulletCount = bulletCount;
 		this.bulletSpeed = speed;
 	}
-	public void fireBullet(int x, int y, double direction)
+	
+	
+	/**
+	 * Fires a bullet from this weapon that travels across the screen
+	 * @param x the x coordinate of this weapon
+	 * @param y the y coordinate of this weapon
+	 * @param direction the direction of this weapon
+	 * @post adds a bullet into this weapon
+	 */
+	public void fireBullet(double x, double y, double direction)
 	{
 		bullets.add(new Bullet(x,y,direction,bulletSpeed,getDamage(),true));
 	}
+	
+	/**
+	 * Returns the Bullet objects in this Gun
+	 * @return the list of all Bullet s in this Gun
+	 */
 	public ArrayList<Bullet> getExistingBullets()
 	{
 		return bullets;
